Adapt to OpenSSL 1.1.0
OpenSSL 1.1.0 hid structure internals and provided methods for reading
and writing them. This patch adapts to the changes so that it's
possible to build it against the new as well as as old OpenSSL
library.

Because the new OpenSSL does not support setting each prime number
separately, this patch fakes somes otherwise undefined values. This
looks ugly but this cannot be done better.

I recommend to add new Perl subroutines for setting the prime numbers in a
bulk as new OpenSSL does. These will be a strightforward binding with
less code and higher performance.

CPAN RT#118346